removed the previous AAC code
