+++ .gitignore
@@ -7,6 +7,7 @@ node_modules
 npm-debug.log
 coverage
 .cache
+.claude
 .eslintcache
 .env
 docs/.obsidian
