+++ src/features/parcel/service/parcel.service.test.js
@@ -2,7 +2,7 @@ import {
   splitParcelId,
   getParcelActionsWithAvailableArea
 } from './parcel.service.js'
-import { getAvailableAreaDataRequirements } from '~/src/features/available-area/availableArea.js'
+import { getAvailableAreaDataRequirements } from '~/src/features/available-area/availableAreaDataRequirements.js'
 import { findMaximumAvailableArea } from '~/src/features/available-area/availableArea.lp.js'
 import { formatExplanationSections } from '~/src/features/available-area/explanations.lp.js'
 import {
@@ -12,7 +12,7 @@ import {
 import { vi } from 'vitest'
 
 // Mock the dependencies
-vi.mock('~/src/features/available-area/availableArea.js')
+vi.mock('~/src/features/available-area/availableAreaDataRequirements.js')
 vi.mock('~/src/features/available-area/availableArea.lp.js')
 vi.mock('~/src/features/available-area/explanations.lp.js')
 vi.mock('~/src/features/parcel/transformers/parcelActions.transformer.js')
